third_party/blink/web_tests/wpt_internal/atomic-move-auto/modal-dialog.html

<!DOCTYPE html>
<title>moveBefore should not close a modal dialog</title>
<meta name="variant" content="?action=append">
<meta name="variant" content="?action=prepend">
<meta name="variant" content="?action=before">
<meta name="variant" content="?action=after">
<script src="resources/helper.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<section id="old_parent">
    <dialog id="dialog">
    </dialog>
</section>
<section id="new_parent">
</section>
<script>
promise_test(async t => {
    const dialog = document.querySelector("#dialog");
    dialog.showModal();
    reparent(document.querySelector("#new_parent"), dialog);
    assert_equals(document.elementFromPoint(0, 0), dialog);
}, "when reparenting a modal dialog, the dialog should stay modal");
</script>
